+++ Common/src/main/java/com/hypherionmc/sdlink/api/messaging/discord/DiscordMessage.java
@@ -161,7 +161,7 @@ public final class DiscordMessage {
                 String content = this.messageType == MessageType.CHAT ?
                         SDLinkConfig.INSTANCE.messageFormatting.chat
                                 .replace("%player%", author.getDisplayName())
-                                .replace("%mcname%", author.getProfile() == null ? "Unknown" : author.getProfile().getName().replace("_", "\\_"))
+                                .replace("%mcname%", author.getProfile() == null ? "Unknown" : author.getProfile().getName())
                                 .replace("%message%", message)
                         : message;
                 builder.setContent(content);
@@ -252,14 +252,14 @@ public final class DiscordMessage {
         }
 
         embedJson = embedJson
-                .replace("%author%", StringEscapeUtils.escapeJson(this.author.getDisplayName().replace("_", "\\_")))
+                .replace("%author%", StringEscapeUtils.escapeJson(this.author.getDisplayName().replace("\\_", "_")))
                 .replace("%avatar%", this.author.getAvatar())
                 .replace("%message_contents%", StringEscapeUtils.escapeJson(this.message))
                 .replace("%player_avatar%", this.author.getRealPlayerAvatar())
                 .replace("%role_color%", String.valueOf(this.author.getColor()))
-                .replace("%player_name%", StringEscapeUtils.escapeJson(this.author.getRealPlayerName().replace("_", "\\_")))
+                .replace("%player_name%", StringEscapeUtils.escapeJson(this.author.getRealPlayerName().replace("\\_", "_")))
                 .replace("%current_time%", String.valueOf(Instant.now().getEpochSecond()))
-                .replace("%username%", StringEscapeUtils.escapeJson(this.author.getUsername().replace("_", "\\_")));
+                .replace("%username%", StringEscapeUtils.escapeJson(this.author.getUsername().replace("\\_", "_")));
 
         DiscordEmbed embed = EmbedManager.gson.fromJson(embedJson, DiscordEmbed.class);
         return fromData(embed);